Add files via upload
<html>
<head>
<title>My First WebPage</title>
</head>
<body>
<h1><p align="center"><b><u> Gaming </u> </p> </b></h1>

<h2> Advantages of Gaming </h2>

<p> * Gaming provide a fun and social form of entertainment.<br><br>

* Video games benefits include promoting healthy lifestyles, increasing social activity,<br>
  slow the aging process, and make the participant a better decision-maker.<br><br>

* There are plenty of positive effects of video games. Gaming can help those suffering<br>
 from addictions or cravings to decrease the intensity of their desires. They can also<br>
 potentially help those who have multiple sclerosis with balance and improved cognitive function.<br><br>

*The positive effects of video games is in the way they fit into different facets of our daily lives.<br>
 Some of the benefits are physical, psychological, and social. Contrary to popular belief, they can<br>
 promote healthy living and increased social activity through various avenues.<br><br>

*Gaming encourage teamwork and cooperation when played with others.</p><br>

<h2> Disadvantages of Gaming </h2>
<p> *Things like your eyesight. If you are really close to a screen while you are playing, you may begin to notice<br>
 that your vision is more blurry than it was before you started playing. <br>
Keep in mind, though that this would be a long term effect of gaming.<br>
 However, if you are glued to a screen for a few hours and then try to look at <br>
something else, your vision might be blurry, but then as you focus it will return to your regular vision.<br>
 This, added up, could lead to you needing glasses or contacts to help you.<br><br>

*Another common disadvantage is time,
As you become more and more engrossed in gaming, you may lose track of time.<br>
This could cause problems in the long run with your social life. Friends that may not <br>
be as interested in gaming as you are might wonder why you don't spend as much time <br>
with them (this usually only becomes a big problem if you spend a LOT of time gaming).<br>
 Your family may not like that you don't spend more time with them, either. It becomes worse <br>
if you have a family/kids to take care of. If you spend the majority of your time gaming, you aren't<br> 
spending the time with your kids or significant other that they need/like.<br><br>

*Video games can harm your eyesight if you don’t take care, but that is a disadvantage of the medium<br>
 they are presented at (monitors), not the games itselves.</p>



<h6><p align="right"> This page is done by Aathif of Class 7A of Loyola School Made by HTML</h6><br><br><br><br><br><br><br><br><br>


</body>
</html>